gpu: collect the routines a body calls in one place, not two
The offload pass and the device partition each carried a walker that
answered "which routines does this body call", and the two bodies were
identical but for one line: the partition also counts a routine that is
merely named, because whoever is handed it can call it, while the offload
pass wants only what the body actually calls, since that is what it has to
splice.

One collector in gpu_utils.h now, with that one difference as the question
it is asked.
